requirements.txt
absl-py==2.1.0
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
contourpy==1.2.1
cron-descriptor==1.4.3
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.2.0
Django==4.0.2
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.13.0
django-heroku==0.3.1
django-imagekit==4.1.0
django-jazzmin==2.4.9
django-mssql-backend==2.8.1
django-seed==0.3.1
django-storages==1.14.3
django-timezone-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.2.2
executing==2.0.1
Faker==26.0.0
fastjsonschema==2.20.0
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.4.0
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.65.1
h5py==3.11.0
idna==3.7
ipykernel==6.29.5
ipython==8.23.0
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.13.1
keras-pickle-wrapper==1.0.5
kiwisolver==1.4.5
kombu==5.3.7
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.7
notebook==6.0.3
numpy==1.22.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
pandas==1.5.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickle-mixin==1.0.2
pickleshare==0.7.5
pilkit==2.0
Pillow==9.0.1
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
PyJWT==2.8.0
pyodbc==5.1.0
pyparsing==3.1.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.1
rsa==4.9
Send2Trash==1.8.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.4.0
terminado==0.18.1
tinycss2==1.3.0
toposort==1.10
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
whitenoise==6.7.0
wrapt==1.16.0